# docs/swagger.yaml
basePath: /
definitions:
  helpers.CommonErrorResponse:
    properties:
      error:
        example: Bad Request
        type: string
      message: {}
      statusCode:
        example: 400
        type: integer
    type: object
  helpers.CommonResponse:
    properties:
      data: {}
      message:
        example: Success
        type: string
      meta:
        $ref: '#/definitions/helpers.Meta'
      statusCode:
        example: 200
        type: integer
    type: object
  helpers.LoginInput:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: secret123
        type: string
    type: object
  helpers.Meta:
    properties:
      currentPage:
        example: 1
        type: integer
      from:
        example: 1
        type: integer
      lastPage:
        example: 5
        type: integer
      limit:
        example: 10
        type: integer
      to:
        example: 10
        type: integer
      total:
        example: 42
        type: integer
    type: object
  helpers.RegisterInput:
    properties:
      bio:
        example: A passionate reader
        type: string
      email:
        example: john@example.com
        type: string
      password:
        example: secret123
        type: string
      username:
        example: johndoe
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@pubcore.io
    name: API Support
  description: Publishing platform API for managing Books, Authors, and Publishers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: PubCore API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return a JWT token
      parameters:
      - description: User Login Input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/helpers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.CommonErrorResponse'
      summary: Login with email and password
      tags:
      - Auth
  /auth/logout:
    post:
      description: Invalidate the session (client-side token removal)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/helpers.CommonErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout the current user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User Registration Input
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/helpers.RegisterInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.CommonErrorResponse'
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"